requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
imap-tools>=1.4.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-dateutil>=2.8.2
